Extract formatDate from inquiriesTable
